service-unit:refactor: move ingress adapter to primary adapter & move to infrastructure
