+++ service-unit/stateless/internal/application/core/initialization/telemetry.go
@@ -1,7 +1,7 @@
 package initialization
 
 import (
-	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure_new/adapters/secondary/config"
+	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure/adapters/secondary/config"
 	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure/telemetry/tracing"
 	"github.com/hanapedia/the-bench/the-bench-operator/pkg/logger"
 )
