+++ service-unit/stateless/internal/infrastructure/adapters/primary/consumer/kafka/consumer.go
