+++ service-unit/stateless/internal/infrastructure/adapters/secondary/invocation/rest/client.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure_new/adapters/secondary/config"
+	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure/adapters/secondary/config"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 )
