+++ service-unit/stateless/internal/infrastructure/adapters/secondary/invocation/rest/read.go
