+++ service-unit/stateless/internal/infrastructure/adapters/secondary/invocation/rest/write.go
