+++ service-unit/stateless/internal/infrastructure/adapters/secondary/producer/kafka/connection.go
