+++ service-unit/stateless/internal/infrastructure/adapters/secondary/repository/mongo/write.go
